Split server and client rule
