+++ rule/cxf/src/main/java/io/opentracing/contrib/specialagent/rule/cxf/CxfRsServerAgentRule.java
@@ -27,36 +27,22 @@ import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.dynamic.DynamicType.Builder;
 import net.bytebuddy.utility.JavaModule;
 
-public class CxfWsAgentRule extends AgentRule {
+public class CxfRsServerAgentRule extends AgentRule {
   @Override
   public Iterable<? extends AgentBuilder> buildAgent(final AgentBuilder builder) throws Exception {
     return Arrays.asList(
-      builder.type(hasSuperType(named("org.apache.cxf.frontend.ClientFactoryBean")))
+      builder.type(hasSuperType(named("org.apache.cxf.jaxrs.JAXRSServerFactoryBean")))
         .transform(new Transformer() {
           @Override
           public Builder<?> transform(final Builder<?> builder, final TypeDescription typeDescription, final ClassLoader classLoader, final JavaModule module) {
-            return builder.visit(Advice.to(CxfWsClientAdvice.class).on(named("create")));
-          }}),
-      builder.type(hasSuperType(named("org.apache.cxf.frontend.ServerFactoryBean")))
-        .transform(new Transformer() {
-          @Override
-          public Builder<?> transform(final Builder<?> builder, final TypeDescription typeDescription, final ClassLoader classLoader, final JavaModule module) {
-            return builder.visit(Advice.to(CxfWsServerAdvice.class).on(named("create")));
+            return builder.visit(Advice.to(CxfRsServerAdvice.class).on(named("create")));
           }}));
   }
 
-  public static class CxfWsClientAdvice {
+  public static class CxfRsServerAdvice {
     @Advice.OnMethodEnter
     public static void enter(final @Advice.Origin String origin, final @Advice.This Object thiz) {
-      if (isEnabled(CxfWsAgentRule.class.getName(), origin))
-        CxfAgentIntercept.addClientTracingFeature(thiz);
-    }
-  }
-
-  public static class CxfWsServerAdvice {
-    @Advice.OnMethodEnter
-    public static void enter(final @Advice.Origin String origin, final @Advice.This Object thiz) {
-      if (isEnabled(CxfWsAgentRule.class.getName(), origin))
+      if (isEnabled(CxfRsServerAgentRule.class.getName(), origin))
         CxfAgentIntercept.addServerTracingFeauture(thiz);
     }
   }
